+++ tests/test_cli.py
@@ -204,6 +204,7 @@ class ProcessIssuesTest(unittest.TestCase):
                 types.IssueType.Entropy, "foo", types.Chunk("foo", "/bar", {})
             )
         ]
+        mock_scanner.return_value._issue_count = 1  # pylint: disable=protected-access
         runner = CliRunner()
         with runner.isolated_filesystem():
             result = runner.invoke(cli.main, ["scan-local-repo", "."])
@@ -216,6 +217,7 @@ class ProcessIssuesTest(unittest.TestCase):
         self, mock_scanner: mock.MagicMock
     ):
         mock_scanner.return_value.issues = []
+        mock_scanner.return_value.issue_count = 0
         runner = CliRunner()
         with runner.isolated_filesystem():
             result = runner.invoke(cli.main, ["scan-local-repo", "."])
@@ -228,6 +230,7 @@ class ProcessIssuesTest(unittest.TestCase):
         self, mock_scanner: mock.MagicMock
     ):
         mock_scanner.return_value.issues = []
+        mock_scanner.return_value.issue_count = 0
         runner = CliRunner()
         with runner.isolated_filesystem():
             result = runner.invoke(cli.main, ["-q", "-v", "scan-local-repo", "."])
@@ -240,6 +243,7 @@ class ProcessIssuesTest(unittest.TestCase):
         self, mock_scanner: mock.MagicMock
     ):
         mock_scanner.return_value.issues = []
+        mock_scanner.return_value.issue_count = 0
         runner = CliRunner()
         with runner.isolated_filesystem():
             result = runner.invoke(cli.main, ["-q", "scan-local-repo", "."])
@@ -252,6 +256,7 @@ class ProcessIssuesTest(unittest.TestCase):
         self, mock_scanner: mock.MagicMock
     ):
         mock_scanner.return_value.issues = []
+        mock_scanner.return_value.issue_count = 0
         runner = CliRunner()
         with runner.isolated_filesystem():
             result = runner.invoke(cli.main, ["-v", "scan-local-repo", "."])
@@ -267,6 +272,7 @@ class LoggingTests(unittest.TestCase):
         self, mock_formatter: mock.MagicMock, mock_scanner: mock.MagicMock
     ):
         mock_scanner.return_value.issues = []
+        mock_scanner.return_value.issue_count = 0
         runner = CliRunner()
         with runner.isolated_filesystem():
             runner.invoke(cli.main, ["scan-local-repo", "."])
@@ -282,6 +288,7 @@ class LoggingTests(unittest.TestCase):
         self, mock_formatter: mock.MagicMock, mock_scanner: mock.MagicMock
     ):
         mock_scanner.return_value.issues = []
+        mock_scanner.return_value.issue_count = 0
         runner = CliRunner()
         with runner.isolated_filesystem():
             runner.invoke(cli.main, ["--no-log-timestamps", "scan-local-repo", "."])
@@ -295,6 +302,7 @@ class LoggingTests(unittest.TestCase):
         self, mock_formatter: mock.MagicMock, mock_scanner: mock.MagicMock
     ):
         mock_scanner.return_value.issues = []
+        mock_scanner.return_value.issue_count = 0
         runner = CliRunner()
         with runner.isolated_filesystem():
             runner.invoke(cli.main, ["-vvvv", "scan-local-repo", "."])
@@ -307,6 +315,7 @@ class LoggingTests(unittest.TestCase):
     @mock.patch("tartufo.commands.scan_local_repo.GitRepoScanner")
     def test_excess_verbosity_does_not_exceed_debug(self, mock_scanner: mock.MagicMock):
         mock_scanner.return_value.issues = []
+        mock_scanner.return_value._issue_count = 0  # pylint: disable=protected-access
         runner = CliRunner()
         with runner.isolated_filesystem():
             runner.invoke(cli.main, ["-vvvvvvvvvvvvvvv", "scan-local-repo", "."])